+++ sdk/resourcemanager/resources/armresources/CHANGELOG.md
@@ -1,10 +1,24 @@
 # Release History
 
-## 1.2.0-beta.1 (2023-06-12)
+## 2.0.0 (2023-07-13)
+### Breaking Changes
+
+- Type of `DeploymentProperties.Parameters` has been changed from `any` to `map[string]*DeploymentParameter`
+- Type of `DeploymentPropertiesExtended.Providers` has been changed from `[]*Provider` to `[]*DeploymentsProvider`
+- Type of `DeploymentWhatIfProperties.Parameters` has been changed from `any` to `map[string]*DeploymentParameter`
+- Function `*DeploymentsClient.NewListByResourceGroupPager` has been removed
+- Operation `*TagsClient.CreateOrUpdateAtScope` has been changed to LRO, use `*TagsClient.BeginCreateOrUpdateAtScope` instead.
+- Operation `*TagsClient.DeleteAtScope` has been changed to LRO, use `*TagsClient.BeginDeleteAtScope` instead.
+- Operation `*TagsClient.UpdateAtScope` has been changed to LRO, use `*TagsClient.BeginUpdateAtScope` instead.
 
 ### Features Added
 
-- Support for test fakes and OpenTelemetry trace spans.
+- New struct `DeploymentParameter`
+- New struct `DeploymentsProvider`
+- New struct `DeploymentsResourceType`
+- New struct `KeyVaultParameterReference`
+- New struct `KeyVaultReference`
+
 
 ## 1.1.1 (2023-04-14)
 ### Bug Fixes